Revert "Sync implydata/druid master branch with apache/druid master branch (#32)" (#34)
This reverts commit f7692551594cecdebe823145e061a05ee7da9093.

+++ processing/src/main/java/org/apache/druid/segment/data/ColumnarFloats.java
@@ -81,7 +81,7 @@ public interface ColumnarFloats extends Closeable
         @Override
         public double getDouble(int offset)
         {
-          return (double) ColumnarFloats.this.get(offset);
+          return ColumnarFloats.this.get(offset);
         }
 
         @Override
